Rename session.py to client.py
