+++ ghga_connector/core/file_operations.py
@@ -29,11 +29,11 @@ from typing import Any, Iterator, Sequence, Tuple, Union
 
 import crypt4gh.keys
 import crypt4gh.lib
-import requests
+import httpx
 
 from ghga_connector.core import exceptions
+from ghga_connector.core.client import HttpxClient
 from ghga_connector.core.constants import TIMEOUT
-from ghga_connector.core.session import HttpxClient
 
 
 class Crypt4GHEncryptor:
@@ -108,7 +108,7 @@ def download_content_range(
         response = HttpxClient.get(
             download_url, headers=headers, timeout=TIMEOUT, allow_redirects=False
         )
-    except requests.exceptions.RequestException as request_error:
+    except httpx.RequestError as request_error:
         exceptions.raise_if_max_retries(request_error=request_error, url=download_url)
         raise exceptions.RequestFailedError(url=download_url) from request_error
 
@@ -202,7 +202,7 @@ def upload_file_part(*, presigned_url: str, part: bytes) -> None:
 
     try:
         response = HttpxClient.put(presigned_url, data=part, timeout=TIMEOUT)
-    except requests.exceptions.RequestException as request_error:
+    except httpx.RequestError as request_error:
         exceptions.raise_if_max_retries(request_error=request_error, url=presigned_url)
         raise exceptions.RequestFailedError(url=presigned_url) from request_error
